march 1 updates (see description)
We need a space in the privacy policy paragraph on the main page: “…opportunities.Read…” → “…opportunities. Read…”

In Privacy policy and the process page, the logo is doing the huge thing. Can we fix this?

Can we remove the clipping of the icon?

Keyboard navigation around the logo displays oddly. Can we find a fix for this?

Lets turn on stylistic set 5 for the centre tagline “varied fonts for varied vision”.

For the Mobile Breakpoint can we hide/ delete/ display: none everything below the about paragraph (No FAQ, Accessibility etc…)